Added Bearer token to authorize api's.

+++ ToDo/main.py
@@ -5,6 +5,7 @@ from db import engine, session_local
 from pydantic import BaseModel, Field
 from typing import Optional
 from sqlalchemy.orm import Session
+from auth import get_current_user, get_user_exception
 
 app = FastAPI()
 
@@ -28,15 +29,27 @@ class ToDo(BaseModel):
 
 
 
-@app.get('/')
+@app.get('')
 async def get_todos(db: Session = Depends(get_db)):
     return db.query(models.ToDo).all()
 
 
+@app.get('/user/todos')
+async def get_all_todos_by_user( user: dict = Depends(get_current_user), 
+                                 db: Session = Depends(get_db)
+                                    ):
+    if user is None:
+        raise get_user_exception()
+    
+    return db.query(models.ToDo).filter(models.ToDo.owner_id == user.get("id")).all()
+
+
 
 @app.get('/todo/{todo_id}')
-async def get_todo_by_id(todo_id: int, db: Session = Depends(get_db)):
-    todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
+async def get_todo_by_id(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).filter(models.ToDo.owner_id == user.get("id")).first()
     if todo:
         return todo
     raise HTTPException(status_code=404, detail="ToDo item not found.")
@@ -44,12 +57,16 @@ async def get_todo_by_id(todo_id: int, db: Session = Depends(get_db)):
 
 
 @app.post('/create-todo')
-async def create_todo(todo: ToDo, db: Session = Depends(get_db)):
+async def create_todo(todo: ToDo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    
     todo_model = models.ToDo()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get('id') 
     db.add(todo_model)
     db.commit()
 
@@ -60,8 +77,12 @@ async def create_todo(todo: ToDo, db: Session = Depends(get_db)):
 
 
 @app.put('/update-todo/{todo_id}')
-async def update_todo(todo_id: int, todo: ToDo, db: Session = Depends(get_db)):
-    todo_to_update = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
+async def update_todo(todo_id: int, todo: ToDo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    
+    if user is None:
+        raise get_user_exception()
+
+    todo_to_update = db.query(models.ToDo).filter(models.ToDo.id == todo_id).filter(models.ToDo.owner_id == user.get("id")).first()
     if not todo_to_update:
         raise HTTPException(status_code=404, detail="ToDo item not found.")
     todo_to_update.title = todo.title 
@@ -79,8 +100,10 @@ async def update_todo(todo_id: int, todo: ToDo, db: Session = Depends(get_db)):
 
 
 @app.delete('/delete-todo/{todo_id}')
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).first()
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception()
+    todo = db.query(models.ToDo).filter(models.ToDo.id == todo_id).filter(models.ToDo.owner_id == user.get("id")).first()
     if not todo:
         raise HTTPException(status_code=404, detail=f"Todo with todo_id {todo_id} not found.")
     db.query(models.ToDo).filter(models.ToDo.id == todo_id).delete()
